HTTPS: Fix HTTP/2 connection routing
RFC7540 has a feature called 'connection coalescing', which Chrome
implements. This means that two backends behind the same IP can be
misrouted if they use the same certificate and we only route on the TCP
layer (the seconde connection will be coalesced and therefore the TCP
rules will not be triggered again). Fix this by decrypting everything
in one section and only after that splitting into backends.